requirements.txt
# general modules
Django==4.2.4
djangorestframework
django-compressor==4.3.1
django-otp==1.1.6
python-decouple
Pillow

# third party modules
sorl-thumbnail==12.9.0
django-ckeditor==6.7.0
django-taggit==4.0.0
django-extensions==3.2.3
django-js-asset==2.1.0
django-robots==5.0
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt


# email third party modules
django-mail-templated

# deployment modules


# refactoring and tests
black
flake8
